requirements.txt
dj-database-url==2.1.0
dj-static==0.0.6
Django==5.0.4
django-allauth==0.59.0
django-cors-headers==4.5.0
django-extensions==3.2.3
django-jazzmin==2.6.0
django-recaptcha==4.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
gunicorn==21.2.0
Markdown==3.7
python-decouple==3.8
wheel==0.42.0
